refactor: MakeGroupFragmentをActivityに変更
