+++ android-project/MakaseteChoice/app/src/main/AndroidManifest.xml
@@ -17,6 +17,7 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+        <activity android:name=".makegroup.MakeGroupActivity"/>
     </application>
 
 </manifest>
