+++ android-project/MakaseteChoice/app/src/main/java/com/teambrack/makasetechoice/di/ApplicationComponent.kt
@@ -2,6 +2,7 @@ package com.teambrack.makasetechoice.di
 
 import com.teambrack.makasetechoice.MakaseteChoiceApplication
 import com.teambrack.makasetechoice.di.modules.MainActivityModule
+import com.teambrack.makasetechoice.di.modules.MakeGroupActivityModule
 import com.teambrack.makasetechoice.di.modules.RepositoryModule
 import com.teambrack.makasetechoice.di.modules.ViewModelModule
 import dagger.BindsInstance
@@ -15,6 +16,7 @@ import javax.inject.Singleton
     modules = [
         AndroidSupportInjectionModule::class,
         MainActivityModule::class,
+        MakeGroupActivityModule::class,
         RepositoryModule::class,
         ViewModelModule::class
     ]
